Make `BloodPressureSummaryView` implement `BloodPressureSummaryUi`
