+++ app/src/main/java/org/simple/clinic/summary/bloodpressures/BloodPressureSummaryView.kt
@@ -8,6 +8,8 @@ import androidx.cardview.widget.CardView
 import kotlinx.android.synthetic.main.patientsummary_bpsummary_content.view.*
 import org.simple.clinic.R
 import org.simple.clinic.bp.BloodPressureMeasurement
+import org.simple.clinic.di.injector
+import org.simple.clinic.summary.PatientSummaryConfig
 import org.simple.clinic.util.RelativeTimestampGenerator
 import org.simple.clinic.util.UserClock
 import org.simple.clinic.util.UtcClock
@@ -15,6 +17,8 @@ import org.threeten.bp.Duration
 import org.threeten.bp.Instant
 import org.threeten.bp.ZoneId
 import org.threeten.bp.format.DateTimeFormatter
+import javax.inject.Inject
+import javax.inject.Named
 
 private typealias EditMeasurementClicked = (BloodPressureMeasurement) -> Unit
 private typealias NewBpClicked = () -> Unit
@@ -22,7 +26,28 @@ private typealias NewBpClicked = () -> Unit
 class BloodPressureSummaryView(
     context: Context,
     attributeSet: AttributeSet
-) : CardView(context, attributeSet) {
+) : CardView(context, attributeSet), BloodPressureSummaryUi {
+
+  @Inject
+  lateinit var utcClock: UtcClock
+
+  @Inject
+  lateinit var config: PatientSummaryConfig
+
+  @Inject
+  lateinit var timestampGenerator: RelativeTimestampGenerator
+
+  @field:[Inject Named("exact_date")]
+  lateinit var exactDateFormatter: DateTimeFormatter
+
+  @field:[Inject Named("time_for_bps_recorded")]
+  lateinit var timeFormatterForBp: DateTimeFormatter
+
+  @Inject
+  lateinit var zoneId: ZoneId
+
+  @Inject
+  lateinit var userClock: UserClock
 
   init {
     LayoutInflater.from(context).inflate(R.layout.patientsummary_bpsummary_content, this, true)
@@ -31,7 +56,30 @@ class BloodPressureSummaryView(
   var editMeasurementClicked: EditMeasurementClicked? = null
   var newBpClicked: NewBpClicked? = null
 
-  fun render(
+  override fun onFinishInflate() {
+    super.onFinishInflate()
+    if(isInEditMode) {
+      return
+    }
+
+    context.injector<BloodPressureSummaryViewInjector>().inject(this)
+  }
+
+  override fun populateBloodPressures(bloodPressureMeasurements: List<BloodPressureMeasurement>) {
+    render(
+        bloodPressureMeasurements = bloodPressureMeasurements,
+        utcClock = utcClock,
+        placeholderLimit = config.numberOfBpPlaceholders,
+        timestampGenerator = timestampGenerator,
+        dateFormatter = exactDateFormatter,
+        canEditFor = config.bpEditableDuration,
+        bpTimeFormatter = timeFormatterForBp,
+        zoneId = zoneId,
+        userClock = userClock
+    )
+  }
+
+  private fun render(
       bloodPressureMeasurements: List<BloodPressureMeasurement>,
       utcClock: UtcClock,
       placeholderLimit: Int,
