+++ app/src/main/java/org/simple/clinic/summary/bloodpressures/BloodPressureSummaryViewInjector.kt
@@ -1,0 +1,5 @@
+package org.simple.clinic.summary.bloodpressures
+
+interface BloodPressureSummaryViewInjector {
+  fun inject(target: BloodPressureSummaryView)
+}
